+++ CHANGELOG.rst
@@ -3,8 +3,8 @@ Changelog
 
 All notable changes to this project will be documented in this file.
 
-[Unreleased]
-------------
+[2.0.0] - 2020-10-05
+--------------------
 
 Added
 ^^^^^
@@ -12,7 +12,8 @@ Added
 * Training library (Multi-stream Kalman filter training).
 * Tools for HDNNP data set handling, etc.
 * Python interface (basic functionality).
-* Sphinx documentation.
+* Sphinx documentation (+ Doxygen API reference).
+* CabanaMD interface (by Saaketh Desai and Sam Reeve)
 
 Changed
 ^^^^^^^